enable threaded processing
`Dir.chdir` is not thread safe, so all threads will end up evaluating
inside the same working directory. Instead we can pass the path down and
make sure to use it everywhere. `Open3` accepts a `chdir` argument that
it passes to `Process.spawn` under the hood, which is thread safe (at
least for our purposes).